requirements.txt
# Core requirements for Elite RAT
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
cryptography>=41.0.0
pyyaml>=6.0
pyjwt>=2.8.0
pillow>=10.0.0
dnspython>=2.4.0
psutil>=5.9.0
requests>=2.31.0
python-engineio>=4.5.0
python-socketio>=5.9.0

# Optional but recommended
python-dotenv>=1.0.0
colorama>=0.4.6